main.go
package main

import (
	"context"
	"fmt"
	"net/http"
	"os"
	"os/signal"

	chi "github.com/go-chi/chi/v5"
	middleware "github.com/go-chi/chi/v5/middleware"
	redoc "github.com/go-openapi/runtime/middleware"
	"github.com/identitatem/idp-configs-api/config"
	l "github.com/identitatem/idp-configs-api/logger"
	"github.com/prometheus/client_golang/prometheus/promhttp"
	log "github.com/sirupsen/logrus"
)

func setupDocsMiddleware(handler http.Handler) http.Handler {
	opt := redoc.RedocOpts{
		SpecURL: "/api/idp-configs-api/v0/openapi.json",
	}
	return redoc.Redoc(opt, handler)
}

// Handler function that responds with Hello World
func helloWorld(w http.ResponseWriter, r *http.Request) {
	fmt.Fprint(w, "Hello world")
}

// statusOK returns a simple 200 status code
func statusOK(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "text/plain; charset=utf-8")
	w.WriteHeader(http.StatusOK)
}

// Serve OpenAPI spec json
func serveOpenAPISpec(w http.ResponseWriter, r *http.Request) {	
	cfg := config.Get()
	http.ServeFile(w, r, cfg.OpenAPIFilePath)
}

func initDependencies() {
	config.Init()
	l.InitLogger()
	// db.InitDB()
}

func main() {
	initDependencies()
	cfg := config.Get()
	log.WithFields(log.Fields{
		"Hostname":         cfg.Hostname,
		"Auth":             cfg.Auth,
		"WebPort":          cfg.WebPort,
		"MetricsPort":      cfg.MetricsPort,
		"LogLevel":         cfg.LogLevel,
		"Debug":            cfg.Debug,
		"OpenAPIFilePath ": cfg.OpenAPIFilePath,
	}).Info("Configuration Values:")

	// Initialize router
	r := chi.NewRouter()

	r.Use(
		middleware.Logger,
		setupDocsMiddleware,
	)

	// Register handler functions on server routes

	// Health check
	r.Get("/", statusOK)

	// Hello World endpoint
	r.Get("/api/idp-configs-api/v0/ping", helloWorld)

	// OpenAPI Spec
	r.Get("/api/idp-configs-api/v0/openapi.json", serveOpenAPISpec)

	// Router for metrics
	mr := chi.NewRouter()
	mr.Get("/", statusOK)
	mr.Handle("/metrics", promhttp.Handler())

	srv := http.Server{
		Addr:    fmt.Sprintf(":%d", cfg.WebPort),
		Handler: r,
	}

	msrv := http.Server{
		Addr:    fmt.Sprintf(":%d", cfg.MetricsPort),
		Handler: mr,
	}	

	idleConnsClosed := make(chan struct{})
	go func() {
		sigint := make(chan os.Signal, 1)
		signal.Notify(sigint, os.Interrupt)
		<-sigint
		if err := srv.Shutdown(context.Background()); err != nil {
			log.WithFields(log.Fields{"error": err}).Fatal("HTTP Server Shutdown failed")
		}
		if err := msrv.Shutdown(context.Background()); err != nil {
			log.WithFields(log.Fields{"error": err}).Fatal("HTTP Server Shutdown failed")
		}
		close(idleConnsClosed)
	}()
	
	go func() {
		if err := msrv.ListenAndServe(); err != http.ErrServerClosed {
			log.WithFields(log.Fields{"error": err}).Fatal("Metrics Service Stopped")
		}
	}()

	if err := srv.ListenAndServe(); err != http.ErrServerClosed {
		log.WithFields(log.Fields{"error": err}).Fatal("Service Stopped")
	}

	<-idleConnsClosed
	log.Info("Everything has shut down, goodbye")
}
